refactor(#741): refactor authController
